fix(DatePicker): fixing types in DatePicker stories
